+++ Packages/com.unity.inputsystem/InputSystem/Plugins/Switch/NPad.cs
@@ -1,4 +1,4 @@
-#if UNITY_EDITOR || UNITY_SWITCH
+#if UNITY_EDITOR || UNITY_SWITCH || UNITY_STANDALONE_OSX || UNITY_STANDALONE_WIN || UNITY_WSA
 using System;
 using System.Runtime.InteropServices;
 using UnityEngine.InputSystem.Controls;
@@ -11,12 +11,13 @@ using UnityEngine.InputSystem.Utilities;
 
 namespace UnityEngine.InputSystem.Switch.LowLevel
 {
+#if UNITY_EDITOR || UNITY_SWITCH
     /// <summary>
-    /// Structure of HID input reports for Switch NPad controllers.
+    /// Structure of native input reports for Switch NPad controllers.
     /// </summary>
     /// <seealso href="http://en-americas-support.nintendo.com/app/answers/detail/a_id/22634/~/joy-con-controller-diagram"/>
     [StructLayout(LayoutKind.Explicit, Size = 60)]
-    public struct NPadInputState : IInputStateTypeInfo
+    public struct NPadInputStateSwitch : IInputStateTypeInfo
     {
         public FourCC format
         {
@@ -101,7 +102,7 @@ namespace UnityEngine.InputSystem.Switch.LowLevel
             A = East,
         }
 
-        public NPadInputState WithButton(Button button, bool value = true)
+        public NPadInputStateSwitch WithButton(Button button, bool value = true)
         {
             var bit = (uint)1 << (int)button;
             if (value)
@@ -126,13 +127,13 @@ namespace UnityEngine.InputSystem.Switch.LowLevel
         public InputDeviceCommand baseCommand;
 
         [FieldOffset(InputDeviceCommand.kBaseCommandSize + 0)]
-        public NPad.NpadId npadId;
+        public NPadSwitch.NpadId npadId;
         [FieldOffset(InputDeviceCommand.kBaseCommandSize + 1)]
-        public NPad.Orientation orientation;
+        public NPadSwitch.Orientation orientation;
         [FieldOffset(InputDeviceCommand.kBaseCommandSize + 2)]
         public short padding0;
         [FieldOffset(InputDeviceCommand.kBaseCommandSize + 4)]
-        public NPad.NpadStyles styleMask;
+        public NPadSwitch.NpadStyles styleMask;
         [FieldOffset(InputDeviceCommand.kBaseCommandSize + 8)]
         public int colorLeftMain;
         [FieldOffset(InputDeviceCommand.kBaseCommandSize + 12)]
@@ -228,14 +229,14 @@ namespace UnityEngine.InputSystem.Switch.LowLevel
         public InputDeviceCommand baseCommand;
 
         [FieldOffset(InputDeviceCommand.kBaseCommandSize + 0)]
-        public NPad.Orientation orientation;
+        public NPadSwitch.Orientation orientation;
 
         public FourCC typeStatic
         {
             get { return Type; }
         }
 
-        public static NpadDeviceIOCTLSetOrientation Create(NPad.Orientation orientation)
+        public static NpadDeviceIOCTLSetOrientation Create(NPadSwitch.Orientation orientation)
         {
             return new NpadDeviceIOCTLSetOrientation
             {
@@ -343,16 +344,112 @@ namespace UnityEngine.InputSystem.Switch.LowLevel
             };
         }
     }
+#endif
+
+#if UNITY_EDITOR || UNITY_STANDALONE_OSX || UNITY_STANDALONE_WIN || UNITY_WSA
+    /// <summary>
+    /// Structure of HID input reports for Switch Pro controllers.
+    /// </summary>
+    [StructLayout(LayoutKind.Explicit, Size = 20)]
+    public struct NPadInputStateHID : IInputStateTypeInfo
+    {
+        public FourCC format => new FourCC('H', 'I', 'D');
+
+        [InputControl(name = "dpad", format = "BIT", layout = "Dpad",  bit = 24, sizeInBits = 4, defaultState = 8)]
+        [InputControl(name = "dpad/up", format = "BIT", layout = "DiscreteButton", parameters = "minValue=7,maxValue=1,nullValue=8,wrapAtValue=7", bit = 24, sizeInBits = 4)]
+        [InputControl(name = "dpad/right", format = "BIT", layout = "DiscreteButton", parameters = "minValue=1,maxValue=3", bit = 24, sizeInBits = 4)]
+        [InputControl(name = "dpad/down", format = "BIT", layout = "DiscreteButton", parameters = "minValue=3,maxValue=5", bit = 24, sizeInBits = 4)]
+        [InputControl(name = "dpad/left", format = "BIT", layout = "DiscreteButton", parameters = "minValue=5, maxValue=7", bit = 24, sizeInBits = 4)]
+        [InputControl(name = "buttonNorth", displayName = "X", bit = (uint)Button.North)]
+        [InputControl(name = "buttonSouth", displayName = "B", bit = (uint)Button.South, usage = "Back")]
+        [InputControl(name = "buttonWest", displayName = "Y", bit = (uint)Button.West, usage = "SecondaryAction")]
+        [InputControl(name = "buttonEast", displayName = "A", bit = (uint)Button.East, usage = "PrimaryAction")]
+        [InputControl(name = "leftStickPress", displayName = "Left Stick", bit = (uint)Button.StickL)]
+        [InputControl(name = "rightStickPress", displayName = "Right Stick", bit = (uint)Button.StickR)]
+        [InputControl(name = "leftShoulder", displayName = "L", bit = (uint)Button.L)]
+        [InputControl(name = "rightShoulder", displayName = "R", bit = (uint)Button.R)]
+        [InputControl(name = "leftTrigger", displayName = "ZL", format = "BIT", bit = (uint)Button.ZL)]
+        [InputControl(name = "rightTrigger", displayName = "ZR", format = "BIT", bit = (uint)Button.ZR)]
+        [InputControl(name = "start", displayName = "Plus", bit = (uint)Button.Plus, usage = "Menu")]
+        [InputControl(name = "select", displayName = "Minus", bit = (uint)Button.Minus)]
+        [FieldOffset(0)]
+        public uint buttons;
+
+        [InputControl(name = "leftStick", format = "VC2S", layout = "Stick")]
+        [InputControl(name = "leftStick/x", offset = 0, format = "USHT", parameters = "normalize,normalizeMin=0.15,normalizeMax=0.85,normalizeZero=0.5")]
+        [InputControl(name = "leftStick/left", offset = 0, format = "USHT", parameters = "normalize,normalizeMin=0.15,normalizeMax=0.85,normalizeZero=0.5,clamp,clampMin=0.15,clampMax=0.5,invert")]
+        [InputControl(name = "leftStick/right", offset = 0, format = "USHT", parameters = "normalize,normalizeMin=0.15,normalizeMax=0.85,normalizeZero=0.5,clamp,clampMin=0.5,clampMax=0.85")]
+        [InputControl(name = "leftStick/y", offset = 2, format = "USHT", parameters = "invert,normalize,normalizeMin=0.15,normalizeMax=0.85,normalizeZero=0.5")]
+        [InputControl(name = "leftStick/up", offset = 2, format = "USHT", parameters = "normalize,normalizeMin=0.15,normalizeMax=0.85,normalizeZero=0.5,clamp,clampMin=0.15,clampMax=0.5,invert")]
+        [InputControl(name = "leftStick/down", offset = 2, format = "USHT", parameters = "normalize,normalizeMin=0.15,normalizeMax=0.85,normalizeZero=0.5,clamp,clampMin=0.5,clampMax=0.85,invert=false")]
+        [FieldOffset(4)] public ushort leftStickX;
+        [FieldOffset(6)] public ushort leftStickY;
+
+        [InputControl(name = "rightStick", format = "VC2S", layout = "Stick")]
+        [InputControl(name = "rightStick/x", offset = 0, format = "USHT", parameters = "normalize,normalizeMin=0.15,normalizeMax=0.85,normalizeZero=0.5")]
+        [InputControl(name = "rightStick/left", offset = 0, format = "USHT", parameters = "normalize,normalizeMin=0.15,normalizeMax=0.85,normalizeZero=0.5,clamp,clampMin=0,clampMax=0.5,invert")]
+        [InputControl(name = "rightStick/right", offset = 0, format = "USHT", parameters = "normalize,normalizeMin=0.15,normalizeMax=0.85,normalizeZero=0.5,clamp,clampMin=0.5,clampMax=1")]
+        [InputControl(name = "rightStick/y", offset = 2, format = "USHT", parameters = "invert,normalize,normalizeMin=0.15,normalizeMax=0.85,normalizeZero=0.5")]
+        [InputControl(name = "rightStick/up", offset = 2, format = "USHT", parameters = "normalize,normalizeMin=0.15,normalizeMax=0.85,normalizeZero=0.5,clamp,clampMin=0.15,clampMax=0.5,invert")]
+        [InputControl(name = "rightStick/down", offset = 2, format = "USHT", parameters = "normalize,normalizeMin=0.15,normalizeMax=0.85,normalizeZero=0.5,clamp,clampMin=0.5,clampMax=0.85,invert=false")]
+        [FieldOffset(8)] public ushort rightStickX;
+        [FieldOffset(10)] public ushort rightStickY;
+
+        public float leftTrigger => ((buttons & (1 << (int)Button.ZL)) != 0) ? 1f : 0f;
+
+        public float rightTrigger => ((buttons & (1 << (int)Button.ZR)) != 0) ? 1f : 0f;
+
+        public enum Button
+        {
+            North = 11,
+            South = 8,
+            West = 10,
+            East = 9,
+
+            StickL = 18,
+            StickR = 19,
+            L = 12,
+            R = 13,
+
+            ZL = 14,
+            ZR = 15,
+            Plus = 17,
+            Minus = 16,
+
+            X = North,
+            B = South,
+            Y = West,
+            A = East,
+        }
+
+        public NPadInputStateHID WithButton(Button button, bool value = true)
+        {
+            var bit = (uint)1 << (int)button;
+            if (value)
+                buttons |= bit;
+            else
+                buttons &= ~bit;
+            // dpad default state
+            buttons |= 8 << 24;
+            leftStickX = 0x8000;
+            leftStickY = 0x8000;
+            rightStickX = 0x8000;
+            rightStickY = 0x8000;
+            return this;
+        }
+    }
+#endif
 }
 
 namespace UnityEngine.InputSystem.Switch
 {
+#if UNITY_EDITOR || UNITY_SWITCH
     /// <summary>
     /// An NPad controller for Switch, which can be a Joy-Con.
     /// </summary>
     /// <seealso cref="NPadInputState"/>
-    [InputControlLayout(stateType = typeof(NPadInputState), displayName = "Switch Controller (on Switch)")]
-    public class NPad : Gamepad, INPadRumble
+    [InputControlLayout(stateType = typeof(NPadInputStateSwitch), displayName = "Switch Controller (on Switch)")]
+    public class NPadSwitch : Gamepad, INPadRumble
     {
         public ButtonControl leftSL { get; private set; }
         public ButtonControl leftSR { get; private set; }
@@ -662,5 +759,13 @@ namespace UnityEngine.InputSystem.Switch
             ExecuteCommand(ref cmd);
         }
     }
+#endif
+
+#if UNITY_EDITOR || UNITY_STANDALONE_OSX || UNITY_STANDALONE_WIN || UNITY_WSA
+    [InputControlLayout(stateType = typeof(NPadInputStateHID), displayName = "Switch Controller (on HID)")]
+    public class NPadHID : Gamepad
+    {
+    }
+#endif
 }
 #endif // UNITY_EDITOR || UNITY_SWITCH