+++ Packages/com.unity.inputsystem/InputSystem/Plugins/Switch/SwitchSupport.cs
@@ -1,4 +1,4 @@
-#if UNITY_EDITOR || UNITY_SWITCH
+#if UNITY_EDITOR || UNITY_SWITCH || UNITY_STANDALONE_OSX || UNITY_STANDALONE_WIN || UNITY_WSA
 using UnityEngine.InputSystem.Layouts;
 
 namespace UnityEngine.InputSystem.Switch
@@ -15,11 +15,20 @@ namespace UnityEngine.InputSystem.Switch
     {
         public static void Initialize()
         {
-            InputSystem.RegisterLayout<NPad>(
+        #if UNITY_EDITOR || UNITY_SWITCH
+            InputSystem.RegisterLayout<NPadSwitch>(
                 matches: new InputDeviceMatcher()
                     .WithInterface("Switch")
                     .WithManufacturer("Nintendo")
                     .WithProduct("Wireless Controller"));
+        #endif
+        #if UNITY_EDITOR || UNITY_STANDALONE_OSX || UNITY_STANDALONE_WIN || UNITY_WSA
+            InputSystem.RegisterLayout<NPadHID>(
+                matches: new InputDeviceMatcher()
+                    .WithInterface("HID")
+                    .WithCapability("vendorId", 0x57e) // Nintendo
+                    .WithCapability("productId", 0x2009)); // Pro Controller.
+        #endif
         }
     }
 }