+++ Packages/com.unity.inputsystem/CHANGELOG.md
@@ -20,6 +20,7 @@ however, it has to be formatted properly to pass verification tests.
 ### Added
 
 - Added support for generic joysticks on WebGL (which don't use the standard gamepad mapping).
+- Added support for Nintendo Switch Pro Controllers on desktops.
 
 ### Changed